Split RadialBasis to separate file
