+++ src/atomic/TwoDBasis.h
@@ -20,7 +20,7 @@
 #include "../general/polynomial_basis.h"
 #include "../general/model_potential.h"
 #include "../general/sap.h"
-#include "basis.h"
+#include "RadialBasis.h"
 
 namespace helfem {
   namespace atomic {
